move attack module selector component to redteaming client folder
