+++ app/views/redteaming/components/attackModuleSelector.tsx
